style: change upgrader version date issue #5084
