+++ src/scene_server/admin_server/imports.go
@@ -121,6 +121,6 @@ import (
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202102261105"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202103031533"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202103231621"
-	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202103241156"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202104011012"
+	_ "configcenter/src/scene_server/admin_server/upgrader/y3.9.202104121004"
 )
