+++ src/scene_server/admin_server/upgrader/y3.9.202104121004/pkg.go
@@ -10,42 +10,37 @@
  * limitations under the License.
  */
 
-package y3_9_202103241156
+package y3_9_202104121004
 
 import (
 	"context"
 
-	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/scene_server/admin_server/upgrader"
 	"configcenter/src/storage/dal"
-	"configcenter/src/storage/dal/types"
 )
 
-func instanceObjectIDMapping(ctx context.Context, db dal.RDB, conf *upgrader.Config) (err error) {
+func init() {
+	upgrader.RegistUpgrader("y3.9.202104121004", upgrade)
+}
 
-	exists, err := db.HasTable(ctx, objectBaseMapping)
-	if err != nil {
-		blog.ErrorJSON("check table(%s) exist error, err: %s", objectBaseMapping, err)
+func upgrade(ctx context.Context, db dal.RDB, conf *upgrader.Config) (err error) {
+	blog.Info("y3.9.202104121004")
+
+	if err := instanceObjectIDMapping(ctx, db, conf); err != nil {
+		blog.Errorf("[upgrade y3.9.202104121004] migrate instance object id mapping table failed, error:%s",
+			err.Error())
 		return err
-
-	}
-	if !exists {
-		if err := db.CreateTable(ctx, objectBaseMapping); err != nil {
-			blog.ErrorJSON("create table(%s) error, err: %s", objectBaseMapping, err)
-			return err
-		}
-
 	}
 
-	index := types.Index{
-		Name: common.CCLogicIndexNamePrefix + "InstID",
-		Keys: map[string]int32{
-			common.BKInstIDField: 1,
-		},
-		Background: true,
+	err = splitTable(ctx, db, conf)
+	if err != nil {
+		blog.Errorf("[upgrade y3.9.202104121004] migrate inst split table failed, error  %s", err.Error())
+		return err
 	}
-	if err := db.Table(objectBaseMapping).CreateIndex(ctx, index); err != nil && !db.IsDuplicatedError(err) {
+
+	if err = syncInnerObjectIndex(ctx, db, conf); err != nil {
+		blog.Errorf("[upgrade y3.9.202104121004] migrate inner object index failed, error  %s", err.Error())
 		return err
 	}
