+++ src/scene_server/admin_server/upgrader/y3.9.202104121004/split_inst_table.go
@@ -10,7 +10,7 @@
  * limitations under the License.
  */
 
-package y3_9_202103241156
+package y3_9_202104121004
 
 import (
 	"context"
